+++ Firmware/FFBoard/UserExtensions/Inc/FFBHIDMain.h
@@ -35,7 +35,7 @@
 
 class FFBHIDMain: public FFBoardMain, public cpp_freertos::Thread, PersistentStorage,ExtiHandler,public UsbHidHandler, ErrorHandler{
 	enum class FFBWheel_commands : uint32_t{
-		ffbactive,axes,btntypes,lsbtn,addbtn,aintypes,lsain,addain,hidrate,hidsendspd,estop
+		ffbactive,axes,btntypes,lsbtn,addbtn,aintypes,lsain,addain,hidrate,hidsendspd,estop,cfrate
 	};
 
 public: