+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Added min/max commands for apin
 - Added analog processing functions to ADS111X
 - Selecting a "none" encoder will remove the axis value. Allows analog inputs to be used as the primary axis.
+- Added constant force rate command